refactoring
Tuners, Terminators and Resampling objects are now created and used once
per call. Main problem was with `bt = tnr('mbo')` when used sequentially
with two AutoTuner's which had learners with tuning spaces that required
different surrogates in MBO. For example, the first tuning space had only
numerics so a GP was used as a surrogate and that model remained in the
`bt` object and was used later on a learner with a tuning space that had
a factor, which GP doesn't support, causing the MBO to fail.